Basic auth based templating

+++ basic/controllers/PilotController.php
@@ -6,8 +6,11 @@ use app\models\Country;
 use app\models\Pilot;
 use app\models\PilotSearch;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException ;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use Yii;
+
 
 /**
  * PilotController implements the CRUD actions for Pilot model.
@@ -113,22 +116,26 @@ class PilotController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Pilot();
+        if(Yii::$app->user->can('userCrud')){
+            $model = new Pilot();
 
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
-                return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post()) && $model->save()) {
+                    return $this->redirect(['view', 'id' => $model->id]);
+                }
+            } else {
+                $model->loadDefaultValues();
             }
+
+            $countries = Country::find()->select(['name'])->indexBy('id')->column();
+
+            return $this->render('create', [
+                'model' => $model,
+                'countries' => $countries,
+            ]);
         } else {
-            $model->loadDefaultValues();
+            throw new ForbiddenHttpException();
         }
-
-        $countries = Country::find()->select(['name'])->indexBy('id')->column();
-
-        return $this->render('create', [
-            'model' => $model,
-            'countries' => $countries,
-        ]);
     }
 
     /**
@@ -140,15 +147,19 @@ class PilotController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(Yii::$app->user->can('userCrud')){
+            $model = $this->findModel($id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+        } else {
+            throw new ForbiddenHttpException();
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -160,9 +171,13 @@ class PilotController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(Yii::$app->user->can('userCrud')){
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+            throw new ForbiddenHttpException();
+        }
     }
 
     /**
